+++ pkg/services/grafana-apiserver/storage/file/util.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Kubernetes Authors.
 
-package json
+package file
 
 import (
 	"bytes"
